[Refactor]: Style of Code Writing
Remove unwanted additional 'keysSeen' object. Relied on the accumulator
object for checking if a value is already existing

Changed 'nonOverlappingKey' to isOverlappingKey as its easier to read
and understand the purpose. Further, the check process to see if a key
is already existing in the code is simplified a lot more.

Seperated validation and overlapping checks into a seperate folder
for cleaner code and presentation

Avoided the 'MirrarrayError' class as its not required in the code.